+++ perun/logic/config.py
@@ -167,6 +167,7 @@ format:
     shortlog: "%checksum:6% (%stats%) %desc%"
     output_profile_template: "%collector%-%cmd%-%args%-%workload%-%date%"
     output_show_template: "%collector%-%cmd%-%args%-%workload%-%date%"
+    sort_profiles_by: time
 
 degradation:
     apply: all